+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/usages/TraitMultinameUsage.java
@@ -18,6 +18,7 @@ package com.jpexs.decompiler.flash.abc.usages;
 
 import com.jpexs.decompiler.flash.abc.ABC;
 import com.jpexs.decompiler.flash.abc.types.traits.Traits;
+import java.util.Objects;
 
 /**
  *
@@ -25,19 +26,23 @@ import com.jpexs.decompiler.flash.abc.types.traits.Traits;
  */
 public abstract class TraitMultinameUsage extends MultinameUsage implements InsideClassMultinameUsageInterface {
 
-    public int traitIndex;
+    protected final int traitIndex;
 
     public static final int TRAITS_TYPE_CLASS = 1;
+
     public static final int TRAITS_TYPE_INSTANCE = 2;
+
     public static final int TRAITS_TYPE_SCRIPT = 3;
 
-    public int traitsType;
-    public int classIndex;
-    public int scriptIndex;
+    protected final int traitsType;
 
-    public Traits traits;
+    protected final int classIndex;
 
-    public int parentTraitIndex;
+    protected final int scriptIndex;
+
+    protected final Traits traits;
+
+    protected final int parentTraitIndex;
 
     public TraitMultinameUsage(ABC abc, int multinameIndex, int scriptIndex, int classIndex, int traitIndex, int traitsType, Traits traits, int parentTraitIndex) {
         super(abc, multinameIndex);
@@ -55,6 +60,54 @@ public abstract class TraitMultinameUsage extends MultinameUsage implements Insi
     }
 
     @Override
+    public int hashCode() {
+        int hash = super.hashCode();
+        hash = 23 * hash + this.traitIndex;
+        hash = 23 * hash + this.traitsType;
+        hash = 23 * hash + this.classIndex;
+        hash = 23 * hash + this.scriptIndex;
+        hash = 23 * hash + Objects.hashCode(this.traits);
+        hash = 23 * hash + this.parentTraitIndex;
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        if (!super.equals(obj)) {
+            return false;
+        }
+        final TraitMultinameUsage other = (TraitMultinameUsage) obj;
+        if (this.traitIndex != other.traitIndex) {
+            return false;
+        }
+        if (this.traitsType != other.traitsType) {
+            return false;
+        }
+        if (this.classIndex != other.classIndex) {
+            return false;
+        }
+        if (this.scriptIndex != other.scriptIndex) {
+            return false;
+        }
+        if (this.parentTraitIndex != other.parentTraitIndex) {
+            return false;
+        }
+        if (!Objects.equals(this.traits, other.traits)) {
+            return false;
+        }
+        return true;
+    }
+
+    @Override
     public boolean collides(MultinameUsage other) {
         return false;
     }
@@ -64,4 +117,23 @@ public abstract class TraitMultinameUsage extends MultinameUsage implements Insi
         return classIndex;
     }
 
+    public int getTraitIndex() {
+        return traitIndex;
+    }
+
+    public int getTraitsType() {
+        return traitsType;
+    }
+
+    public int getScriptIndex() {
+        return scriptIndex;
+    }
+
+    public Traits getTraits() {
+        return traits;
+    }
+
+    public int getParentTraitIndex() {
+        return parentTraitIndex;
+    }
 }